+++ src/Lotto.js
@@ -1,28 +1,13 @@
-const { ERROR_MESSAGES } = require('./constant/messages');
+const Validator = require('./validator');
 
 class Lotto {
   #numbers;
 
+  validator = new Validator();
+
   constructor(numbers) {
-    this.validate(numbers);
+    this.validator.validateLotto(numbers);
     this.#numbers = numbers;
-  }
-
-  validate(numbers) {
-    this.validateLength(numbers);
-    this.validateDoubled(numbers);
-  }
-
-  validateLength(numbers) {
-    if (numbers.length !== 6) {
-      throw new Error(ERROR_MESSAGES.NUMBERS_LENGTH_MUST_BE_SIX);
-    }
-  }
-
-  validateDoubled(numbers) {
-    const set = Array.from(new Set(numbers));
-    if (numbers.length !== set.length)
-      throw new Error(ERROR_MESSAGES.NUMBERS_MUST_NOT_OVERLAP);
   }
 
   getNumbers() {
